# contract/openapi.json
{
  "components": {
    "schemas": {
      "ActivityCategory": {
        "description": "The ONLY valid category values anywhere in the system (CONTRACTS.md\n\u00a72, \u00a78) \u2014 budget-breakdown grouping, seed data, and frontend filters\nall depend on this exact set. Add a value here first if it's ever\nneeded, never ad hoc in one layer.",
        "enum": [
          "sightseeing",
          "food",
          "adventure",
          "transport",
          "stay",
          "other"
        ],
        "title": "ActivityCategory",
        "type": "string"
      },
...
      "ActivityRead": {
        "properties": {
          "category": {
            "$ref": "#/components/schemas/ActivityCategory"
          },
          "city_id": {
            "format": "uuid",
            "title": "City Id",
            "type": "string"
          },
          "cost": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cost"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "duration_hours": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Hours"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "image_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Url"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "city_id",
          "category"
        ],
        "title": "ActivityRead",
        "type": "object"
      },
...
      "CityRead": {
        "properties": {
          "cost_index": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cost Index"
          },
          "country": {
            "title": "Country",
            "type": "string"
          },
          "google_place_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Google Place Id"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "image_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Url"
          },
          "latitude": {
            "title": "Latitude",
            "type": "number"
          },
          "longitude": {
            "title": "Longitude",
            "type": "number"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "popularity": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Popularity"
          }
        },
        "required": [
          "id",
          "name",
          "country",
          "latitude",
          "longitude"
        ],
        "title": "CityRead",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ItineraryActivityRead": {
        "properties": {
          "activity": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ActivityRead"
              },
              {
                "type": "null"
              }
            ]
          },
          "activity_id": {
            "format": "uuid",
            "title": "Activity Id",
            "type": "string"
          },
          "cost": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cost"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "scheduled_date": {
            "format": "date",
            "title": "Scheduled Date",
            "type": "string"
          },
          "scheduled_time": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scheduled Time"
          },
          "stop_id": {
            "format": "uuid",
            "title": "Stop Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "stop_id",
          "activity_id",
          "scheduled_date"
        ],
        "title": "ItineraryActivityRead",
        "type": "object"
      },
...
      "PublicTripRead": {
        "description": "Read-only trip served by GET /public/trips/{token}.\n\nDeliberately NOT TripRead: this response carries no `id`, no `user_id`,\nand no `share_token`. A public viewer gets the itinerary, never the\ninternal ids that would let them probe authenticated routes, and never\na token they could reshare beyond the one they already hold.",
        "properties": {
          "amount_total": {
            "title": "Amount Total",
            "type": "string"
          },
          "cover_image_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cover Image Url"
          },
          "date_end": {
            "format": "date",
            "title": "Date End",
            "type": "string"
          },
          "date_start": {
            "format": "date",
            "title": "Date Start",
            "type": "string"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "owner_name": {
            "title": "Owner Name",
            "type": "string"
          },
          "stops": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/StopRead"
            },
            "title": "Stops",
            "type": "array"
          }
        },
        "required": [
          "name",
          "date_start",
          "date_end",
          "owner_name",
          "amount_total"
        ],
        "title": "PublicTripRead",
        "type": "object"
      },
...
      "StopRead": {
        "description": "Stop as returned by the API.\n\nThe last five fields are computed per request by services/feasibility.py\n(CONTRACTS \u00a77.2) \u2014 they are NOT columns. The frontend displays them and\nnever recomputes any of them itself (CONTRACTS \u00a78).",
        "properties": {
          "city": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CityRead"
              },
              {
                "type": "null"
              }
            ]
          },
          "city_id": {
            "format": "uuid",
            "title": "City Id",
            "type": "string"
          },
          "date_end": {
            "format": "date",
            "title": "Date End",
            "type": "string"
          },
          "date_start": {
            "format": "date",
            "title": "Date Start",
            "type": "string"
          },
          "distance_from_previous_km": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Distance From Previous Km"
          },
          "distance_source": {
            "anyOf": [
              {
                "enum": [
                  "directions",
                  "haversine"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Distance Source"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "is_feasible": {
            "default": true,
            "title": "Is Feasible",
            "type": "boolean"
          },
          "itinerary_activities": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/ItineraryActivityRead"
            },
            "title": "Itinerary Activities",
            "type": "array"
          },
          "sequence": {
            "title": "Sequence",
            "type": "integer"
          },
          "travel_duration_hours": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Travel Duration Hours"
          },
          "travel_gap_days": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Travel Gap Days"
          },
          "trip_id": {
            "format": "uuid",
            "title": "Trip Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "trip_id",
          "city_id",
          "sequence",
          "date_start",
          "date_end"
        ],
        "title": "StopRead",
        "type": "object"
      },
...
      "TripCreate": {
        "properties": {
          "budget_target": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Budget Target"
          },
          "cover_image_url": {
            "anyOf": [
              {
                "maxLength": 1024,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cover Image Url"
          },
          "date_end": {
            "format": "date",
            "title": "Date End",
            "type": "string"
          },
          "date_start": {
            "format": "date",
            "title": "Date Start",
            "type": "string"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "name": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name",
          "date_start",
          "date_end"
        ],
        "title": "TripCreate",
        "type": "object"
      },
      "TripDetailRead": {
        "description": "A trip plus its ordered stops \u2014 what the Itinerary Builder/View loads.",
        "properties": {
          "budget_target": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Budget Target"
          },
          "cover_image_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cover Image Url"
          },
          "date_end": {
            "format": "date",
            "title": "Date End",
            "type": "string"
          },
          "date_start": {
            "format": "date",
            "title": "Date Start",
            "type": "string"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "is_public": {
            "title": "Is Public",
            "type": "boolean"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "share_token": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Share Token"
          },
          "stops": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/StopRead"
            },
            "title": "Stops",
            "type": "array"
          },
          "user_id": {
            "format": "uuid",
            "title": "User Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "date_start",
          "date_end",
          "user_id",
          "is_public"
        ],
        "title": "TripDetailRead",
        "type": "object"
      },
      "TripRead": {
        "properties": {
          "budget_target": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Budget Target"
          },
          "cover_image_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cover Image Url"
          },
          "date_end": {
            "format": "date",
            "title": "Date End",
            "type": "string"
          },
          "date_start": {
            "format": "date",
            "title": "Date Start",
            "type": "string"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "is_public": {
            "title": "Is Public",
            "type": "boolean"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "share_token": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Share Token"
          },
          "user_id": {
            "format": "uuid",
            "title": "User Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "date_start",
          "date_end",
          "user_id",
          "is_public"
        ],
        "title": "TripRead",
        "type": "object"
      },
      "TripUpdate": {
        "description": "PATCH /trips/{id}.\n\n`is_public` is here, but `share_token` deliberately is NOT: the token is\nserver-generated on first share and never regenerated implicitly\n(CONTRACTS \u00a72), so it is not a client-settable field.",
        "properties": {
          "budget_target": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Budget Target"
          },
          "cover_image_url": {
            "anyOf": [
              {
                "maxLength": 1024,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cover Image Url"
          },
          "date_end": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Date End"
          },
          "date_start": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Date Start"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "is_public": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Public"
          },
          "name": {
            "anyOf": [
              {
                "maxLength": 255,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          }
        },
        "title": "TripUpdate",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "auth/login"
          }
        },
        "type": "oauth2"
      }
    }
  },
  "info": {
    "title": "GlobeTrotter API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/public/trips/{token}": {
      "get": {
        "description": "Read-only itinerary for anyone holding the link. No write actions exposed.",
        "operationId": "read_public_trip_public_trips__token__get",
        "parameters": [
          {
            "in": "path",
            "name": "token",
            "required": true,
            "schema": {
              "title": "Token",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicTripRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Read Public Trip",
        "tags": [
          "public"
        ]
      }
    },
...
    "/trips": {
      "get": {
        "description": "Trips owned by OR shared with the current user (CONTRACTS \u00a74), most\nrecent departure first.",
        "operationId": "list_trips_trips_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/TripRead"
                  },
                  "title": "Response List Trips Trips Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "List Trips",
        "tags": [
          "trips"
        ]
      },
      "post": {
        "operationId": "create_trip_trips_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TripCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TripRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Create Trip",
        "tags": [
          "trips"
        ]
      }
    },
    "/trips/{trip_id}": {
      "delete": {
        "operationId": "delete_trip_trips__trip_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "trip_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Trip Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Delete Trip",
        "tags": [
          "trips"
        ]
      },
      "get": {
        "description": "A trip plus its ordered stops, with the \u00a77.2 computed fields.",
        "operationId": "read_trip_trips__trip_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "trip_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Trip Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TripDetailRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Read Trip",
        "tags": [
          "trips"
        ]
      },
      "patch": {
        "description": "Update a trip.\n\nToggling `is_public` is owner-only (CONTRACTS \u00a75) and mints the share\ntoken on first publish. The token is never regenerated on later\npublishes (CONTRACTS \u00a72), so a link already shared with someone keeps\nworking after a private/public round trip.",
        "operationId": "update_trip_trips__trip_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "trip_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Trip Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TripUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TripRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Update Trip",
        "tags": [
          "trips"
        ]
      }
    },
...
  }
}